Extract a GradleVersion API
